sanjaya/requirements_deploy.txt
fastapi
uvicorn
xgboost
shap
pandas
scikit-learn
joblib
boto3
requests
python-dotenv
langchain
langgraph
psycopg2-binary
sqlalchemy
python-multipart
openai
pyjwt
cryptography
passlib[bcrypt]
python-jose[cryptography]
python-json-logger
prometheus-client
jaeger-client
# OAuth2 & MFA
python-jose
authlib
google-auth-oauthlib
